cosmetic changes and added login page
