+++ app/src/main/AndroidManifest.xml
@@ -11,11 +11,12 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SettingsActivity"></activity>
+        <activity android:name=".SettingsActivity"/>
         <activity android:name=".AddActivity" />
         <activity android:name=".NavigationActivity" />
         <activity android:name=".ProfileActivity" />
         <activity android:name=".SearchActivity" />
+        <activity android:name=".HomeActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
