+++ app/src/main/java/com/example/cs407project/HomeActivity.java
@@ -1,0 +1,58 @@
+package com.example.cs407project;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.MenuItem;
+
+import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+public class HomeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_home);
+
+        BottomNavigationView navigation = findViewById(R.id.bottom_navigation);
+        navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
+
+        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.fragment_map);
+    }
+
+    private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener =
+            new BottomNavigationView.OnNavigationItemSelectedListener() {
+                @Override
+                public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                    switch (item.getItemId()) {
+                        case R.id.action_search:
+                            Intent searchIntent = new Intent(HomeActivity.this, SearchActivity.class);
+                            startActivity(searchIntent);
+                            return true;
+                        case R.id.action_profile:
+                            Intent profileIntent = new Intent(HomeActivity.this, ProfileActivity.class);
+                            startActivity(profileIntent);
+                            return true;
+                        case R.id.action_navigation:
+                            Intent navigationIntent = new Intent(HomeActivity.this, NavigationActivity.class);
+                            startActivity(navigationIntent);
+                            return true;
+                        case R.id.action_add:
+                            Intent addIntent = new Intent(HomeActivity.this, AddActivity.class);
+                            startActivity(addIntent);
+                            return true;
+                        case R.id.action_settings:
+                            Intent settingsIntent = new Intent(HomeActivity.this, SettingsActivity.class);
+                            startActivity(settingsIntent);
+                            return true;
+                        default:
+                            return false;
+                    }
+                }
+
+            };
+
+}
