+++ app/src/main/java/com/example/cs407project/MainActivity.java
@@ -1,58 +1,14 @@
 package com.example.cs407project;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
 import android.os.Bundle;
-import android.view.MenuItem;
 
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.material.bottomnavigation.BottomNavigationView;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        BottomNavigationView navigation = findViewById(R.id.bottom_navigation);
-        navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
-
-        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.fragment_map);
+        setContentView(R.layout.activity_home);
     }
-
-    private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener =
-            new BottomNavigationView.OnNavigationItemSelectedListener() {
-                @Override
-                public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                    switch (item.getItemId()) {
-                        case R.id.action_search:
-                            Intent searchIntent = new Intent(MainActivity.this, SearchActivity.class);
-                            startActivity(searchIntent);
-                            return true;
-                        case R.id.action_profile:
-                            Intent profileIntent = new Intent(MainActivity.this, ProfileActivity.class);
-                            startActivity(profileIntent);
-                            return true;
-                        case R.id.action_navigation:
-                            Intent navigationIntent = new Intent(MainActivity.this, NavigationActivity.class);
-                            startActivity(navigationIntent);
-                            return true;
-                        case R.id.action_add:
-                            Intent addIntent = new Intent(MainActivity.this, AddActivity.class);
-                            startActivity(addIntent);
-                            return true;
-                        case R.id.action_settings:
-                            Intent settingsIntent = new Intent(MainActivity.this, SettingsActivity.class);
-                            startActivity(settingsIntent);
-                            return true;
-                        default:
-                            return false;
-                    }
-                }
-
-            };
-
 }
